requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cloudinary==1.41.0
Django==5.1.3
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-environ==0.11.2
django-js-asset==3.0.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.9.1
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==2.2.3
